refactor: 엔티티 생성 메서드에서 검증 로직 제거
- 엔티티 생성 과정에서 수행하던 유효성 검사를 제거
- Validation Layer에서 검증을 수행하도록 추후 변경 예정
