+++ movie-domain/src/main/java/com/example/domain/model/entity/Movie.java
@@ -1,0 +1,65 @@
+package com.example.domain.model.entity;
+
+import com.example.domain.converter.ContentRatingConverter;
+import com.example.domain.model.base.AuditingFields;
+import com.example.domain.model.valueObject.ContentRating;
+import jakarta.persistence.CascadeType;
+import jakarta.persistence.Column;
+import jakarta.persistence.Convert;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.OneToMany;
+import jakarta.persistence.OrderBy;
+import java.time.LocalDate;
+import java.util.List;
+import lombok.Getter;
+
+@Getter
+@Entity
+public class Movie extends AuditingFields {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(nullable = false) private String title;
+    @Convert(converter = ContentRatingConverter.class)
+    @Column(nullable = false) private ContentRating contentRating;
+    @Column(nullable = false) private LocalDate releaseDate;
+    @Column(nullable = false) private String thumbnailUrl;
+    @Column(nullable = false) private int runtimeMinutes;
+    @Column(nullable = false) private String genre;
+
+    @OneToMany(mappedBy = "movie", cascade = CascadeType.ALL)
+    @OrderBy("startTime ASC")
+    private List<Screening> screenings;
+
+
+    protected Movie() {}
+
+    private Movie(String title,
+                  ContentRating contentRating,
+                  LocalDate releaseDate,
+                  String thumbnailUrl,
+                  int runtimeMinutes,
+                  String genre) {
+        this.title = title;
+        this.contentRating = contentRating;
+        this.releaseDate = releaseDate;
+        this.thumbnailUrl = thumbnailUrl;
+        this.runtimeMinutes = runtimeMinutes;
+        this.genre = genre;
+    }
+
+    public static Movie of(String title,
+                           ContentRating contentRating,
+                           LocalDate releaseDate,
+                           String thumbnailUrl,
+                           int runtimeMinutes,
+                           String genre) {
+        return new Movie(title, contentRating, releaseDate, thumbnailUrl, runtimeMinutes, genre);
+    }
+
+}
